commit 09cd07d8848d7d709bacb1518be29c65058a2dd6 Coroutines with activity scope

+++ MyApplication/app/src/main/java/com/example/myapplication/MediaItem.kt
@@ -1,5 +1,5 @@
 package com.example.myapplication
 
-data class MediaItem(val id: Int, val title: String, val thumbUrl: String, val type: Type) {
+data class MediaItem(val id: Long, val title: String, val thumbUrl: String, val type: Type) {
     enum class Type { PHOTO, VIDEO }
 }
